% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordCorrection.R
\name{.coordCorrection}
\alias{.coordCorrection}
\title{Imports:
stats}
\usage{
.coordCorrection(
  x,
  y,
  rotate,
  mid,
  textdistance = NULL,
  distance,
  outerstart,
  width,
  outside,
  bending,
  discreteAxis = FALSE
)
}
\arguments{
\item{x}{vector, x values of all data points}

\item{y}{vector, y value of all data points}

\item{rotate}{number in degrees. Defines where the brace points to: 0=up (default), 90=right, 180=down, 270=left}

\item{mid}{number from 0.25 to 0.75. Position of the pointer within the brace space. If NULL (default), will be determined by data.}

\item{textdistance}{number. Distance of the label to the brace pointer}

\item{distance}{number. Space between the brace and the nearest data point. If NULL (default), will be determined by data.}

\item{outerstart}{number. If not NULL, overwrites distance and sets all braces to the same origin}

\item{width}{number. Distance from the brace's start to its tip. If NULL (default), will be determined by data.}

\item{outside}{boolean. If TRUE (default), brace is next to the data area. If FALSE, brace is inside the data area}

\item{bending}{number from 0 to 0.5. Determines bend of the brace curves (0=rectangular). If NULL (default), will be determined by data. If too high, values will result in zick-zack lines}

\item{discreteAxis}{boolean. Set to TRUE if the axis along which the brace expands is discrete (often true for bar graphs)}
}
\description{
Imports:
stats
}
