% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekBrace.R
\name{.seekBrace}
\alias{.seekBrace}
\title{Imports:
stats}
\usage{
.seekBrace(x, y, rotate, bending, npoints)
}
\arguments{
\item{x}{vector, x values of all data points}

\item{y}{vector, y value of all data points}

\item{rotate}{number, defines where the brace is pointing to: 0=up, 90=right, 180=down, 270=left. When specified by user, will overwrite other directions the brace might have from x/y coordinates.}

\item{bending}{number, how strongly the curves of the braces should be bent (the higher the more round). Note: too high values will result in the brace showing zick-zack lines}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}
}
\description{
Imports:
stats
}
