% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlyBraces.R
\name{seekBrace}
\alias{seekBrace}
\title{creates a dataframe for a curled brace}
\usage{
seekBrace(
  xstart = 0,
  xend = 1,
  ystart = 0,
  yend = 5,
  mid = 0.5,
  pointing = "side",
  npoints = 100
)
}
\arguments{
\item{xstart}{number, most left part of the brace}

\item{xend}{number, most right part of the brace}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace}

\item{mid}{number, where the pointer is within the bracket space (between 0.25 and 0.75)}

\item{pointing}{string, either "side" or "updown"}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}
}
\value{
data.frame with x and y columns that can be used as an input for ggplot or plotly
}
\description{
creates a dataframe for a curled brace
}
\examples{
library(curlyBrace, ggplot2)
mybrace <- seekBrace()
ggplot() + geom_line(aes(x,y), data=mybrace, orientation="y")

mybrace <- seekBrace(xend=5, yend=1, pointing="updown")
ggplot() + geom_line(aes(x,y), data=mybrace, orientation="x")
}
