% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekBraceLabel.R
\name{seekBraceLabel}
\alias{seekBraceLabel}
\title{dataframe for a curly brace label}
\usage{
seekBraceLabel(
  xstart = 0,
  xend = 1,
  ystart = 0,
  yend = 5,
  mid = 0.5,
  pointing = "updown"
)
}
\arguments{
\item{xstart}{number, most left part of the brace}

\item{xend}{number, most right part of the brace. If pointing="side" the brace points towards this.}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace. If pointing="updown" (default) the brace points towards this.}

\item{mid}{number between 0.25 and 0.75, defining the relative position of the pointer.}

\item{pointing}{string, either "side" or "updown" (default)}
}
\value{
data.frame with columns x and y, to be used as an input for ggplot's geom_text
}
\description{
dataframe for a curly brace label
}
\examples{
library(ggbrace)
library(ggplot2)
mybrace <- seekBrace()
mylabel <- seekBraceLabel()
ggplot(mapping=aes(x,y)) +
  geom_path(data=mybrace) +
  geom_text(data=mylabel, label="my custom brace", vjust=0)

}
