% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_brace.R
\name{stat_brace}
\alias{stat_brace}
\title{create curly braces as a layer in ggplot}
\usage{
stat_brace(
  mapping = NULL,
  data = NULL,
  rotate = 0,
  addlabels = T,
  distance = NULL,
  outerstart = NULL,
  width = NULL,
  mid = NULL,
  npoints = 100,
  textsize = 5,
  ...
)
}
\arguments{
\item{rotate}{integer, either 0, 90, 180 or 270 to indicate if the braces should point up, right, down or left respectively}

\item{addlabels}{boolean, should labels be displayed?}

\item{mid}{number between 0.25 and 0.75, defining the relative position of the pointer.}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace. If pointing="updown" (default) the brace points towards this.}

\item{pointing}{string, either "side" or "updown" (default)}
}
\value{
ggplot2 layer object (geom_path) that can directly be added to a ggplot2 object. If a label was provided, a another layer (geom_text) is added.
}
\description{
create curly braces as a layer in ggplot
}
\examples{
library(ggbrace)
library(ggplot2)
ggplot(mtcars, aes(mpg, wt, color=factor(am))) +
  geom_point() +
  facet_wrap(~vs) +
  stat_brace(rotate=90, aes(label=factor(am)))

}
