% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordCorrection.R
\name{.coordCorrection}
\alias{.coordCorrection}
\title{Imports:
stats}
\usage{
.coordCorrection(
  x,
  y,
  rotate,
  mid,
  textdistance = NULL,
  distance,
  outerstart,
  width,
  outside,
  discreteAxis = FALSE
)
}
\arguments{
\item{x}{vector, x values of all data points}

\item{y}{vector, y value of all data points}

\item{rotate}{number, defines where the brace is pointing to: 0=up, 90=right, 180=down, 270=left. When specified by user, will overwrite other directions the brace might have from x/y coordinates.}

\item{mid}{number, where the pointer is within the bracket space (between 0.25 and 0.75). If NULL (default), will be determined automatically based on the data.}

\item{textdistance}{number, distance of the label to the brace pointer}

\item{distance}{number, space between the brace and the nearest data point}

\item{outerstart}{number, overwrites distance and provides one coordinate for all braces}

\item{width}{number, how wide should the braces be? If NULL (default), will be determined automatically based on the data.}

\item{outside}{boolean, should brace be outside of the data area or cover the data area?}

\item{discreteAxis}{boolean, does the embraced axis feature discrete values (often true for bar graphs)}
}
\description{
Imports:
stats
}
