% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekBrace.R
\name{.seekBrace}
\alias{.seekBrace}
\title{Imports:
stats}
\usage{
.seekBrace(x, y, rotate, bending, npoints, bracketType)
}
\arguments{
\item{x}{vector, x values of all data points}

\item{y}{vector, y value of all data points}

\item{rotate}{number in degrees. Defines where the brace points to: 0=up (default), 90=right, 180=down, 270=left}

\item{bending}{number from 0 to 0.5. Determines bend of the brace curves (0=rectangular). If NULL (default), will be determined by data. If too high, values will result in zick-zack lines}

\item{npoints}{integer. Number of points generated for the brace curves. Will be rounded to be a multiple of 4 for calculation purposes.}

\item{bracketType}{text choice. Either "curly" (default) or "square"}
}
\description{
Imports:
stats
}
