% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekBrace.R
\name{.seekBrace}
\alias{.seekBrace}
\title{Imports:
stats}
\usage{
.seekBrace(x, y, rotate, bending, npoints)
}
\arguments{
\item{x}{vector, x values of all data points}

\item{y}{vector, y value of all data points}

\item{rotate}{number in degrees, defines where the brace points to: 0=up, 90=right, 180=down, 270=left}

\item{bending}{number, how strong is the bend of the brace curves (0=rectangular). If NULL (default), will be determined by data. If too high, values will result in zick-zack lines}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}
}
\description{
Imports:
stats
}
