% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_brace.R
\name{geom_brace}
\alias{geom_brace}
\title{create a curly brace as a layer ready to be added to a ggplot object}
\usage{
geom_brace(
  xstart = 0,
  xend = 1,
  ystart = 0,
  yend = 1,
  mid = 0.5,
  pointing = "updown",
  npoints = 100,
  ...
)
}
\arguments{
\item{xstart}{number, most left part of the brace}

\item{xend}{number, most right part of the brace. If pointing="side" the brace points towards this.}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace. If pointing="updown" (default) the brace points towards this.}

\item{mid}{number, where the pointer is within the bracket space (between 0.25 and 0.75)}

\item{pointing}{string, either "side" or "updown" (default)}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}
}
\value{
ggplot2 layer object (geom_path) that can directly be added to a ggplot2 object
}
\description{
Imports:
ggplot2
}
\examples{
library(curlyBrace, ggplot2)
ggplot() + geom_brace()

ggplot() + geom_brace(ystart=-5, pointing="side")

ggplot() + geom_brace(color="red", size=3, linetype="dashed")
}
