% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_brace.R
\name{geom_brace}
\alias{geom_brace}
\title{create a curly brace as a layer ready to be added to a ggplot object}
\usage{
geom_brace(
  xstart = 0,
  xend = 1,
  ystart = 0,
  yend = 1,
  mid = 0.5,
  pointing = "updown",
  label = NULL,
  labelsize = 5,
  labelcolor = "black",
  npoints = 100,
  ...
)
}
\arguments{
\item{xstart}{number, most left part of the brace}

\item{xend}{number, most right part of the brace. If pointing="side" the brace points towards this.}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace. If pointing="updown" (default) the brace points towards this.}

\item{mid}{number, where the pointer is within the bracket space (between 0.25 and 0.75)}

\item{pointing}{string, either "side" or "updown" (default)}

\item{label}{string, a custom text to be displayed at the brace.}

\item{labelsize}{number, changing the font size of the label. Only takes effect if the parameter "label" was defined.}

\item{labelcolor}{string, defining the color of the label. Only takes effect if the parameter "label" was defined.}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
ggplot2 layer object (geom_path) that can directly be added to a ggplot2 object. If a label was provided, a another layer (annotate) is added.
}
\description{
Imports:
ggplot2
}
\examples{
library(ggbrace)
library(ggplot2)
ggplot() + geom_brace()

ggplot() + geom_brace(ystart=-5, pointing="side")

ggplot() + geom_brace(color="red", size=3, linetype="dashed")
}
