% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekBrace.R
\name{seekBrace}
\alias{seekBrace}
\title{dataframe for a curly brace}
\usage{
seekBrace(
  xstart = 0,
  xend = 1,
  ystart = 0,
  yend = 5,
  mid = 0.5,
  pointing = "updown",
  npoints = 100
)
}
\arguments{
\item{xstart}{number, most left part of the brace}

\item{xend}{number, most right part of the brace. If pointing="side" the brace points towards this.}

\item{ystart}{number, top end of the brace}

\item{yend}{number, bottom end of the brace. If pointing="updown" (default) the brace points towards this.}

\item{mid}{number, where the pointer is within the bracket space (between 0.25 and 0.75)}

\item{pointing}{string, either "side" or "updown" (default)}

\item{npoints}{integer, number of points generated for the brace curves (resolution). This number will be rounded to be a multiple of 4 for calculation purposes.}
}
\value{
data.frame with columns x and y, to be used as an input for ggplot's geom_path
}
\description{
dataframe for a curly brace
}
\examples{
library(ggbrace)
library(ggplot2)
mybrace <- seekBrace()
ggplot() + geom_path(aes(x,y), data=mybrace)

mybrace <- seekBrace(pointing="side")
ggplot() + geom_path(aes(x,y), data=mybrace)
}
